Add new OCR engine kraken
Kraken is an Open Source OCR engine with trainable segmentation and
OCR models. It can work with printed and handwritten texts.

This initial implementation comes with two generic OCR models which
can be used on a wide range of German publications, but also with
other languages which are based on Latin script.

Signed-off-by: Stefan Weil <sw@weilnetz.de>